sonar-batch/src/test/java/org/sonar/batch/scan/filesystem/ExclusionPatternsTest.java
/*
 * Sonar, open source software quality management tool.
 * Copyright (C) 2008-2012 SonarSource
 * mailto:contact AT sonarsource DOT com
 *
 * Sonar is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 3 of the License, or (at your option) any later version.
 *
 * Sonar is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public
 * License along with Sonar; if not, write to the Free Software
 * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02
 */
package org.sonar.batch.scan.filesystem;

import org.junit.Test;
import org.sonar.api.CoreProperties;
import org.sonar.api.config.Settings;
import org.sonar.api.utils.WildcardPattern;

import static org.fest.assertions.Assertions.assertThat;

public class ExclusionPatternsTest {
  @Test
  public void ignore_inclusion_of_world() {
    Settings settings = new Settings();
    settings.setProperty(CoreProperties.PROJECT_INCLUSIONS_PROPERTY, "**/*");
    assertThat(ExclusionPatterns.sourceInclusions(settings)).isEmpty();
  }

  @Test
  public void load_inclusion() {
    Settings settings = new Settings();
    settings.setProperty(CoreProperties.PROJECT_INCLUSIONS_PROPERTY, "**/*Foo.java");
    WildcardPattern[] inclusions = ExclusionPatterns.sourceInclusions(settings);

    assertThat(inclusions).hasSize(1);
    assertThat(inclusions[0].toString()).isEqualTo("**/*Foo.java");
  }
}
